requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
uvicorn[standard]
azure-cognitiveservices-speech==1.44.0
google-cloud-translate==3.15.0
openai==1.35.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.0.1  
aiofiles==24.1.0  

httpx==0.24.1
httpcore==0.17.3
h11==0.14.0
anyio==3.7.1

reportlab==4.0.4
docxtpl==0.20.0
htmldocx==0.0.6